hw16-spring-actuator-data-rest/src/main/java/ru/otus/hw16/service/BookServiceImpl.java
package ru.otus.hw16.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import ru.otus.hw16.model.Book;
import ru.otus.hw16.repository.BookRepository;

@Service
@RequiredArgsConstructor
public class BookServiceImpl implements BookService {

  private final BookRepository bookRepository;

  @Override
  @Transactional
  public Book save(Book entity) {
    return bookRepository.save(entity);
  }
}
